Tracking speed

+++ TrackMe/app/src/main/java/ductranit/me/trackme/models/Session.kt
@@ -11,7 +11,7 @@ class Session {
     @Id
     var id: Long = 0
     var distance: Double = 0.0 // in meter
-    var averageSpeed: Int = 0 // km/h
+    var averageSpeed: Float = 0.0f // m/s
     var startTime: Date? = null
     var endTime: Date? = null
